Added reusable select validator
